requirements.txt
pandas==1.4.1
gensim==3.8.3
fuzzywuzzy==0.18.0
tensorflow==2.7.0
joblib==0.17.0
numpy==1.20.3
spacy==3.2.1
seaborn==0.11.2
streamlit==1.2.0
matplotlib==3.5.0
Pillow==9.0.1
scikit_learn==1.0.2
services==0.1.1
plotly==5.5.0
gunicorn==20.1.0
Flask==2.0.2
